fixx

+++ tasks/orehov_n_jarvis_pass/tests/functional/main.cpp
@@ -51,7 +51,7 @@ class OrehovNJarvisPassFuncTests : public ppc::util::BaseRunFuncTests<InType, Ou
   }
 
   bool CheckTestOutputData(OutType &output_data) final {
-    for (int i = 0; i < output_data.size(); i++) {
+    for (size_t i = 0; i < output_data.size(); i++) {
       std::cout << output_data[i].x << " " << output_data[i].y << "\n";
     }
     return (test_res_ == output_data);
